third_party/blink/web_tests/fast/css/getComputedStyle/getComputedStyle-column-rule.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<script>

description("Tests that the -webkit-column-rule shorthand is computed properly.")

var testContainer = document.createElement("div");
testContainer.contentEditable = true;
document.body.appendChild(testContainer);

testContainer.innerHTML = '<div id="test" style="-webkit-column-rule: thin dotted black;">hello</div>';

e = document.getElementById('test');
computedStyle = window.getComputedStyle(e, null);

shouldBeEqualToString("computedStyle.getPropertyValue('-webkit-column-rule')", '1px dotted rgb(0, 0, 0)');

e.style.webkitColumnRule="10px red";
shouldBeEqualToString("computedStyle.getPropertyValue('-webkit-column-rule')", '10px rgb(255, 0, 0)');

e.style.webkitColumnRuleWidth="medium"
e.style.webkitColumnRuleStyle="dashed"
shouldBeEqualToString("computedStyle.getPropertyValue('-webkit-column-rule')", '3px dashed rgb(255, 0, 0)');

e.style.webkitColumnRule="10px dotted blue"
e.style.webkitColumnRuleStyle="none";
shouldBeEqualToString("computedStyle.getPropertyValue('-webkit-column-rule')", '10px rgb(0, 0, 255)');

e.style.webkitColumnRuleStyle="hidden";
shouldBeEqualToString("computedStyle.getPropertyValue('-webkit-column-rule')", '10px hidden rgb(0, 0, 255)');

document.body.removeChild(testContainer);

</script>
</body>
</html>
