third_party/blink/web_tests/fast/css/getComputedStyle/getComputedStyle-column-rule-expected.txt

Tests that the -webkit-column-rule shorthand is computed properly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS computedStyle.getPropertyValue('-webkit-column-rule') is "1px dotted rgb(0, 0, 0)"
PASS computedStyle.getPropertyValue('-webkit-column-rule') is "10px rgb(255, 0, 0)"
PASS computedStyle.getPropertyValue('-webkit-column-rule') is "3px dashed rgb(255, 0, 0)"
PASS computedStyle.getPropertyValue('-webkit-column-rule') is "10px rgb(0, 0, 255)"
PASS computedStyle.getPropertyValue('-webkit-column-rule') is "10px hidden rgb(0, 0, 255)"
PASS successfullyParsed is true

TEST COMPLETE
